# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  authzapi.createPermProfileBody:
    properties:
      description:
        type: string
      name:
        type: string
      relations:
        items:
          type: string
        type: array
      status:
        type: boolean
    type: object
...
  authzapi.updatePermProfileBody:
    properties:
      description:
        type: string
      name:
        type: string
      orgUnits:
        description: flat string IDs — nil = not provided (keep old), [] = clear,
          ["id"] = set
        items:
          type: string
        type: array
      relations:
        items:
          type: string
        type: array
      resourceGroups:
        items:
          type: string
        type: array
      status:
        type: boolean
    type: object
...
  gmod.ApiErrorResponse:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      details: {}
      message:
        example: Invalid request
        type: string
      status:
        example: false
        type: boolean
    type: object
...
  gmod.PaginatedResponse:
    properties:
      code:
        example: SUCCESS
        type: string
      details: {}
      message:
        example: profiles fetched
        type: string
      pagination:
        $ref: '#/definitions/gmod.Pagination'
      status:
        example: true
        type: boolean
    type: object
  gmod.Pagination:
    properties:
      page:
        type: integer
      perPages:
        type: integer
      sortField:
        type: string
      sortOrder:
        type: string
      totalPages:
        type: integer
      totalRecords:
        type: integer
    type: object
...
  gmod.SuccessDetailResponseAny:
    properties:
      details: {}
      status:
        type: boolean
    type: object
...
  gmod.SuccessMessageResponse:
    properties:
      code:
        example: SUCCESS
        type: string
      message:
        example: Operation successful
        type: string
      status:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: REST API for Gateway system
  title: Gateway API
  version: "1.0"
paths:
...
  /orgs/resource/permissions:
    get:
      parameters:
      - description: page (default 1)
        in: query
        name: page
        type: integer
      - description: items per page (default 10)
        in: query
        name: perPages
        type: integer
      - description: search keyword
        in: query
        name: search
        type: string
      - description: sort field (default createdAt)
        in: query
        name: sortField
        type: string
      - description: 'sort order: asc|desc (default desc)'
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: List resource permission profiles
      tags:
      - ResourcePermissionProfile
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/authzapi.createPermProfileBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gmod.SuccessDetailResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Create resource permission profile
      tags:
      - ResourcePermissionProfile
  /orgs/resource/permissions/{id}:
    delete:
      parameters:
      - description: profileId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.SuccessMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete resource permission profile
      tags:
      - ResourcePermissionProfile
    get:
      parameters:
      - description: profileId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.SuccessDetailResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Get resource permission profile by ID
      tags:
      - ResourcePermissionProfile
    patch:
      consumes:
      - application/json
      parameters:
      - description: profileId
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/authzapi.updatePermProfileBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.SuccessDetailResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Update resource permission profile
      tags:
      - ResourcePermissionProfile
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"